LayoutTests/imported/w3c/web-platform-tests/workers/modules/dedicated-worker-parse-error-failure-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token ';'
CONSOLE MESSAGE: SyntaxError: Unexpected token ';'

FAIL Module worker construction for script with syntax error should dispatch an event named error. assert_equals: expected function "function Event() {
    [native code]
}" but got function "function ErrorEvent() {
    [native code]
}"
FAIL Static import on module worker for script with syntax error should dispatch an event named error. assert_equals: expected function "function Event() {
    [native code]
}" but got function "function ErrorEvent() {
    [native code]
}"
